Doc: add comments
