+++ src/init_stack_data.c
@@ -6,7 +6,7 @@
 /*   By: victofer <victofer@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/30 12:11:28 by victofer          #+#    #+#             */
-/*   Updated: 2023/03/01 13:43:27 by victofer         ###   ########.fr       */
+/*   Updated: 2023/03/01 14:36:47 by victofer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,6 +21,7 @@ void	init_stack(t_stack *stack)
 	i = 0;
 	j = 1;
 	stack->b = (int *)malloc(1 * sizeof(int));
+	stack->temp = (int *)malloc(1 * sizeof(int));
 	stack->a_len = 0;
 	stack->chunk_len = 20;
 	stack->pb_count = 0;
