+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: victofer <victofer@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/16 11:48:14 by victofer          #+#    #+#             */
-/*   Updated: 2023/03/01 13:24:55 by victofer         ###   ########.fr       */
+/*   Updated: 2023/03/01 14:32:12 by victofer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ int	main(int argc, char **argv)
 	char	**nbrs;
 	int		print;
 
-	print = 1;
+	print = 0;
 	if (argc == 1)
 		ft_error();
 	init_stack(&stack);
@@ -43,6 +43,6 @@ int	main(int argc, char **argv)
 	if (print)
 		test_print_stacks(stack);
 	free_stacks(stack);
-	//atexit(show_leaks);
+	atexit(show_leaks);
 	return (0);
 }
